Feat: Refactor code and add directory contents method

+++ src/main/java/com/iposhka/filestorageapi/dto/responce/resourse/FileResponseDto.java
@@ -7,9 +7,9 @@ import lombok.EqualsAndHashCode;
 @EqualsAndHashCode(callSuper = false)
 @Data
 public class FileResponseDto extends ResourceResponseDto {
-    private int size;
+    private long size;
 
-    public FileResponseDto(String path, String name, int size) {
+    public FileResponseDto(String path, String name, long size) {
         this.path = path;
         this.name = name;
         this.size = size;
